Render and place table tests.
Determines if the table test exists, if so removes it, then places the
rendered test under the declaration.

Care has been taken that extra white space is removed, while preserving
comments that are not part of the test function's doc. While generating
the doc for the test function it caps the lines at around 80 characters
without splitting words. Does not add newlines after the test function
if appended to the end of the file.

Main function implemented as the entrypoint from `go generate` and the